CloudgRPC/requirements.txt
grpcio>=1.68.0
grpcio-tools>=1.68.0
protobuf>=5.28.0
python-dotenv>=1.0.0